ethclient,node: refactors docs-gen code to own test file
- also fixup the accidentally-removed workaround
for the cited open-rpc/meta-schema issue; see changes
in node/node.go

Date: 2021-01-21 14:04:37-06:00
