+++ ethclient/docsgen_test.go
@@ -1,0 +1,298 @@
+package ethclient
+
+import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"os"
+	"strings"
+	"testing"
+	"text/template"
+
+	meta_schema "github.com/open-rpc/meta-schema"
+)
+
+// The test in this file is only for use as a development tool.
+
+// TestRPCDiscover_BuildStatic puts the OpenRPC document in build/static/openrpc.json.
+// This is intended to be run as a documentation development tool (as opposed to an actual _test_).
+// NOTE that Go maps don't guarantee order, so the diff between runs can be noisy.
+func TestRPCDiscover_BuildStatic(t *testing.T) {
+	if os.Getenv("COREGETH_GEN_OPENRPC_DOCS") == "" {
+		return
+	}
+	err := os.MkdirAll("../build/static", os.ModePerm)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	backend, _ := newTestBackend(t)
+	client, _ := backend.Attach()
+	defer backend.Close()
+	defer client.Close()
+
+	// Current workaround for https://github.com/open-rpc/meta-schema/issues/356.
+	res, err := backend.InprocDiscovery_DEVELOPMENTONLY().Discover()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Should do it this way.
+	// res := &meta_schema.OpenrpcDocument{}
+	// err = client.Call(res, "rpc.discover")
+	// if err != nil {
+	// 	t.Fatal(err)
+	// }
+
+	data, err := json.MarshalIndent(res, "", "    ")
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = ioutil.WriteFile("../build/static/openrpc.json", data, os.ModePerm)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if res.Methods == nil {
+		return
+	}
+
+	tpl := template.New("openrpc_doc")
+
+	tpl.Funcs(template.FuncMap{
+		// tomap gives a plain map from a JSON representation of a given value.
+		// This is useful because the meta_schema data types, being generated, and conforming to a pretty
+		// complex data type in the first place, are not super fun to interact with directly.
+		"tomap": func(any interface{}) map[string]interface{} {
+			out := make(map[string]interface{})
+			data, _ := json.Marshal(any)
+			json.Unmarshal(data, &out)
+			return out
+		},
+		// asjson returns indented JSON.
+		"asjson": func(any interface{}, prefix, indent int) string {
+			by, _ := json.MarshalIndent(any, strings.Repeat("    ", prefix), strings.Repeat("    ", indent))
+			return string(by)
+		},
+		// bulletJSON handles transforming a JSON JSON schema into bullet points, which I think are more legible.
+		"bulletJSON": printBullet,
+		"sum":        func(a, b int) int { return a + b },
+		// trimNameSpecialChars removes characters that the app-specific content descriptor naming
+		// method will also remove, eg '*hexutil.Uint64' -> 'hexutilUint64'.
+		// These "special" characters were removed because of concerns about by-name arguments
+		// and the use of titles for keys.
+		"trimNameSpecialChars": func(s string) string {
+			remove := []string{".", "*"}
+			for _, r := range remove {
+				s = strings.ReplaceAll(s, r, "")
+			}
+			return s
+		},
+		// methodFormatJSConsole is a pretty-printer that returns the JS console use example for a method.
+		"methodFormatJSConsole": func(m *meta_schema.MethodObject) string {
+			name := string(*m.Name)
+			formattedName := strings.Replace(name, "_", ".", 1)
+			getParamName := func(cd *meta_schema.ContentDescriptorObject) string {
+				if cd.Name != nil {
+					return string(*cd.Name)
+				}
+				return string(*cd.Description)
+			}
+			paramNames := func() (paramNames []string) {
+				if m.Params == nil {
+					return nil
+				}
+				for _, n := range *m.Params {
+					if n.ContentDescriptorObject == nil {
+						continue // Should never happen in our implementation; never uses refs.
+					}
+					paramNames = append(paramNames, getParamName(n.ContentDescriptorObject))
+				}
+				return
+			}()
+			return fmt.Sprintf("%s(%s);", formattedName, strings.Join(paramNames, ","))
+		},
+	})
+
+	tpl, err = tpl.Parse(`
+{{ define "schemaTpl" }}
+	` + "```" + `
+{{ bulletJSON . 1 }}
+	` + "```" + `
+
+	<details class="cite"><summary>View Raw</summary>
+	` + "```" + `
+    {{ asjson . 1 1 }}
+	` + "```" + `
+	</details>
+{{ end }}
+
+{{ define "contentDescTpl" -}}
+{{ $nameyDescription := trimNameSpecialChars .description }}
+{{ if eq .name $nameyDescription }}
+<code>{{ .description }}</code> {{ if .summary }}_{{ .summary }}_{{- end }}
+{{- else -}}
+{{ .name }} <code>{{ .description }}</code> {{ if .summary }}_{{ .summary }}_{{- end }}
+{{- end }}
+
+  + Required: {{ if .required }}✓ Yes{{ else }}No{{- end}}
+{{ if .deprecated }}  + Deprecated: :warning: Yes{{- end}}
+{{ if (or (gt (len .schema) 1) .schema.properties) }} 
+=== "Schema"
+
+	` + "```" + ` Schema
+	{{ bulletJSON .schema 1 }}
+	` + "```" + `
+
+=== "Raw"
+
+	` + "```" + ` Raw
+	{{ asjson .schema 1 1 }}
+	` + "```" + `
+{{ end }}
+{{ end }}
+
+{{ define "methodTpl" }}
+{{ $methodmap := tomap . }}
+### {{ .Name }}
+
+{{ .Summary }}
+
+__Params ({{ .Params | len }})__
+{{ if gt (.Params | len) 0 }}
+{{ if eq $methodmap.paramStructure "by-position" }}Parameters must be given _by position_.{{ else if eq $methodmap.paramStructure "by-name" }}Parameters must be given _by name_.{{ end }}  
+{{ range $index, $param := .Params }}
+{{ $parammap := . | tomap }}
+__{{ sum $index 1 }}:__ {{ template "contentDescTpl" $parammap }}
+{{ end }}
+{{ else }}
+_None_
+{{- end}}
+
+__Result__
+
+{{ if .Result -}}
+{{ $result := .Result | tomap }}
+{{- if ne $result.name "Null" }}
+{{ template "contentDescTpl" $result }}
+{{- else -}}
+_None_
+{{- end }}
+{{- end }}
+
+__Client Method Invocation Examples__
+
+=== "Shell"
+
+	` + "```" + ` shell
+	curl -X POST http://localhost:8545 --data '{"jsonrpc": "2.0", id": 42, "method": "{{ .Name }}", "params": []}'
+	` + "```" + `
+
+=== "Javascript Console"
+
+	` + "```" + ` js
+	{{ methodFormatJSConsole . }}
+	` + "```" + `
+
+{{ $docs := .ExternalDocs | tomap }}
+<details><summary>Source code</summary>
+<p>
+{{ .Description }}
+<a href="{{ $docs.url }}" target="_">View on GitHub →</a>
+</p>
+</details>
+
+---
+{{- end }}
+
+| Entity | Version |
+| --- | --- |
+| Source | <code>{{ .Info.Version }}</code> |
+| OpenRPC | <code>{{ .Openrpc }}</code> |
+
+---
+
+{{ range .Methods }}
+{{ template "methodTpl" . }}
+{{ end }}
+`)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	moduleMethods := func() (grouped map[string][]meta_schema.MethodObject) {
+		if res.Methods == nil {
+			return
+		}
+		grouped = make(map[string][]meta_schema.MethodObject)
+		for _, m := range *res.Methods {
+			moduleName := strings.Split(string(*m.Name), "_")[0]
+			group, ok := grouped[moduleName]
+			if !ok {
+				group = []meta_schema.MethodObject{}
+			}
+			group = append(group, m)
+			grouped[moduleName] = group
+		}
+		return
+	}()
+
+	_ = os.MkdirAll("../docs/JSON-RPC-API/modules", os.ModePerm)
+
+	for module, group := range moduleMethods {
+		fname := fmt.Sprintf("../docs/JSON-RPC-API/modules/%s.md", module)
+		fi, err := os.OpenFile(fname, os.O_CREATE|os.O_RDWR, os.ModePerm)
+		if err != nil {
+			t.Fatal(err)
+		}
+		fi.Truncate(0)
+
+		nDoc := &meta_schema.OpenrpcDocument{}
+		*nDoc = *res
+		nDoc.Methods = (*meta_schema.Methods)(&group)
+		err = tpl.Execute(fi, nDoc)
+		if err != nil {
+			t.Fatal(err)
+		}
+		fi.Sync()
+		fi.Close()
+	}
+}
+
+// printBullet handles transforming a JSON JSON schema into bullet points, which I think are more legible.
+func printBullet(any interface{}, depth int) (out string) {
+	defer func() {
+		out += "\n"
+	}()
+	switch typ := any.(type) {
+	case map[string]interface{}:
+		out += "\n"
+		for k, v := range typ {
+			out += fmt.Sprintf("%s- %s: %s", strings.Repeat("\t", depth), k, printBullet(v, depth+1))
+		}
+	case []interface{}:
+
+		// Don't know why this isn't working. Doesn't fire.
+		// if c, ok := any.([]string); ok {
+		// 	return strings.Join(c, ",")
+		// }
+
+		stringSet := []string{}
+		for _, vv := range typ {
+			if s, ok := vv.(string); ok {
+				stringSet = append(stringSet, s)
+			}
+		}
+		if len(stringSet) == len(typ) {
+			return strings.Join(stringSet, ", ")
+		}
+
+		out += "\n"
+		for _, vv := range typ {
+			out += printBullet(vv, depth+1)
+		}
+	default:
+		return fmt.Sprintf("`%v`", any)
+	}
+	return
+}
